+++ lab-02/src/com/auca/Problem05.java
@@ -12,8 +12,8 @@ public class Problem05 {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         System.out.print("Temperature in degrees Fahrenheit? ");
-        int fahrenheit = scan.nextInt();
-        double celsius = (fahrenheit-32)*(5/9.0);
+        double fahrenheit = scan.nextDouble();
+        double celsius = (fahrenheit-32.0)*(5/9.0);
         System.out.printf("The temperature in degrees Celsius is %.2f\n", celsius);
     }
 }
