+++ lab-02/src/com/auca/Problem06.java
@@ -12,8 +12,8 @@ public class Problem06 {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         System.out.print("Enter a degree in Celsius: ");
-        int celsius = scan.nextInt();
+        double celsius = scan.nextDouble();
         double fahrenheit = (9*celsius)/5.0+32;
-        System.out.printf("%d Celsius is %.1f Fahrenheit\n", celsius, fahrenheit);
+        System.out.printf("%g Celsius is %.1f Fahrenheit\n", celsius, fahrenheit);
     }
 }
