+++ lab-02/src/com/auca/Problem07.java
@@ -13,7 +13,7 @@ public class Problem07 {
         Scanner scan = new Scanner(System.in);
         System.out.print("Enter the radius and length of a cylinder: ");
         double radius = scan.nextDouble();
-        int length = scan.nextInt();
+        double length = scan.nextDouble();
         double area = radius*radius*Math.PI;
         double volume = area*length;
         System.out.printf("The area is %.4f\nThe volume is %.1f\n", area, volume);
